+++ apps/backend/src/modules/user-routes/libs/validation-schemas/validation-schemas.ts
@@ -1,0 +1,5 @@
+export {
+	userRouteCreateValidationSchema,
+	userRouteParametersValidationSchema,
+	userRoutePatchValidationSchema,
+} from "@smartscapes/shared";