+++ apps/frontend/src/modules/routes/libs/types/types.ts
@@ -2,6 +2,7 @@ export { type PatchActionPayload } from "./patch-action-payload.type.js";
 export {
 	type RouteCreateRequestDto,
 	type RouteFindAllOptions,
+	type RouteGetAllItemResponseDto,
 	type RouteGetByIdResponseDto,
 	type RoutePatchRequestDto,
 } from "@smartscapes/shared";