+++ apps/frontend/src/modules/routes/routes.ts
@@ -13,8 +13,13 @@ const routesApi = new RoutesApi({
 export { routesApi };
 export {
 	type RouteCreateRequestDto,
+	type RouteGetAllItemResponseDto,
 	type RouteGetByIdResponseDto,
 	type RoutePatchRequestDto,
 } from "./libs/types/types.js";
 export { routesCreateValidationSchema } from "./libs/validation-schemas/validation-schemas.js";
-export { actions, reducer } from "./slices/routes.js";
+export {
+	actions,
+	routeDetailsReducer,
+	routesReducer,
+} from "./slices/routes.js";