+++ apps/frontend/src/pages/manage-routes/libs/types/row-data.types.ts
@@ -1,7 +1,0 @@
-type RowData = {
-	createdAt: string;
-	id: number;
-	name: string;
-};
-
-export { type RowData };