+++ packages/shared/src/index.ts
@@ -72,15 +72,15 @@ export {
 	PermissionKey,
 } from "./modules/permission/permission.js";
 export {
+	type PointOfInterestPatchRequestDto,
 	pointOfInterestUpdateValidationSchema,
 	PointsOfInterestApiPath,
+	type PointsOfInterestCreateRequestDto,
 	pointsOfInterestCreateValidationSchema,
-	type PointsOfInterestPaginatedResponseDto,
-	type PointsOfInterestPaginatedSummary,
-	type PointsOfInterestQueryRequest,
+	type PointsOfInterestGetAllItemResponseDto,
+	type PointsOfInterestGetAllQuery,
+	type PointsOfInterestGetByIdResponseDto,
 	pointsOfInterestQueryValidationSchema,
-	type PointsOfInterestRequestDto,
-	type PointsOfInterestResponseDto,
 } from "./modules/points-of-interest/points-of-interest.js";
 export {
 	type ReviewCreatePayload,
@@ -109,6 +109,18 @@ export {
 	type UserFollowsRequestDto,
 	type UserUnfollowsParametersDto,
 } from "./modules/user-follows/user-follows.js";
+export {
+	UserRouteApiPath,
+	type UserRouteCreateRequestDto,
+	userRouteCreateValidationSchema,
+	type UserRouteParameters,
+	userRouteParametersValidationSchema,
+	type UserRoutePatchRequestDto,
+	userRoutePatchValidationSchema,
+	type UserRouteResponseDto,
+	UserRouteStatus,
+	type UserRouteStatusType,
+} from "./modules/user-routes/user-routes.js";
 export {
 	type AuthenticatedUserPatchRequestDto,
 	type AuthenticatedUserPatchResponseDto,