+++ packages/shared/src/modules/user-routes/libs/types/user-route-create-request-dto.type.ts
@@ -1,0 +1,5 @@
+type UserRouteCreateRequestDto = {
+	routeId: number;
+};
+
+export { type UserRouteCreateRequestDto };